requirements.txt
aiogram==2.23.1
redis
python-dotenv